requirements.txt
streamlit
tensorflow-cpu
Pillow
requests
python-dotenv